use correct version tag for images
Boilerplate does not allow conditionally assigning the `IMAGE_TAG` variable, but it
*does* allow it for `IMAGE_URI`. Since `IMAGE_TAG` is also being
defined in boilerplate/openshift/osd-container-image/standard.mk:47, the
Boilerplate convention is in-use when `IMAGE_URI` is finally evaluated
and expanded when it's used as part of the build step. Adding the colon
to the assignment in variables.mk:19 tells `make` to evaluate it
immediately when the tag is in the correct format, instead of lazily
evaluating when the build occurs. The immediate evaluation is then
passed on to the conditional assignment.